Reorganize python and xml
