+++ parser/parser/python/handlers/composable_node_container_handler.py
@@ -16,7 +16,7 @@ import ast
 
 from parser.context import ParseContext
 from parser.parser.registry import register_handler
-from parser.parser.utils.common import flatten_once, group_entities_by_type
+from parser.parser.utils import flatten_once, group_entities_by_type
 from parser.resolution.utils import resolve_call_signature
