+++ parser/parser/python/handlers/load_composable_nodes_handler.py
@@ -17,7 +17,7 @@ import ast
 from parser.context import ParseContext
 from parser.parser.postprocessing import simplify_launch_configurations
 from parser.parser.registry import register_handler
-from parser.parser.utils.common import flatten_once, group_entities_by_type
+from parser.parser.utils import flatten_once, group_entities_by_type
 from parser.resolution.utils import resolve_call_signature
